Fix loading of thuliasaur island
Geometry seems missing still

+++ dependencies/zone-utilities/common/eqg_geometry.h
@@ -38,7 +38,7 @@ public:
 	void AddPolygon(Polygon &p) { polys.push_back(p); }
 	void SetName(std::string nname) { name = nname; }
 	
-	std::string GetMaterialName(int32_t idx) { if (idx < 0 || idx >= mats.size()) { return ""; } return mats[idx].GetName(); }
+	std::string GetMaterialName(int32_t idx) { if (idx < 0 || idx >= (int32_t)mats.size()) { return ""; } return mats[idx].GetName(); }
 	std::vector<EQG::Material> &GetMaterials() { return mats; }
 	std::vector<Vertex> &GetVertices() { return verts; }
 	std::vector<Polygon> &GetPolygons() { return polys; }
